Урок 7 | использовали okhttp для синхронного(!) запроса

+++ app/src/main/java/com/gb/myweathersb_gb/model/dto/PartDTO.kt
@@ -17,7 +17,7 @@ data class PartDTO(
     val partName: String,
     val polar: Boolean,
     @SerializedName("prec_mm")
-    val precMm: Int,
+    val precMm: Double,
     @SerializedName("prec_period")
     val precPeriod: Int,
     @SerializedName("prec_prob")
